+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `shellCommands` can take sets with "script", "description", "args" and "show" to generate the shell welcome message for the command.
+
+### Fixed
+- shellCommands does not print any internal setup.
+
 ## [6.1.0] - 2022-06-16
 
 ### Fixed